+++ Package.swift
@@ -18,6 +18,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/simibac/ConfettiSwiftUI", .upToNextMajor(from: "1.0.0")),
+        .package(url: "https://github.com/danielsaidi/OnboardingKit.git", .upToNextMajor(from: "6.0.0")),
         .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "3.4.0")),
         .package(url: "https://github.com/danielsaidi/SystemNotification", .upToNextMajor(from: "0.7.0")),
     ],
@@ -26,6 +27,7 @@ let package = Package(
             name: "KankodaKit",
             dependencies: [
                 "ConfettiSwiftUI",
+                "OnboardingKit",
                 "SwiftUIKit",
                 "SystemNotification"
             ]